+++ components/theme.tsx
@@ -1,6 +1,6 @@
 import { extendTheme } from '@chakra-ui/react'
 
-export const theme = extendTheme({
+const theme = extendTheme({
   styles: {
     global: {
       body: {
@@ -9,3 +9,5 @@ export const theme = extendTheme({
     },
   },
 })
+
+export default theme
