+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import { ChakraProvider } from '@chakra-ui/react'
 import type { AppProps } from 'next/app'
-import { theme } from '@/pages/theme'
+import theme from '@/components/theme'
 import Layout from '@/components/Layout'
 
 export default function App({ Component, pageProps }: AppProps) {
